pristine: fix filename casing
