+++ apps/pristine/src/_components/Model-picker.tsx
@@ -7,7 +7,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "./ui/select";
+} from "./ui/Select";
 
 interface ModelPickerProps {
   selectedModel: modelID;
